Fix warning
git-svn-id: https://svn.osgeo.org/gdal/trunk@25064 f0d54148-0727-0410-94bb-9a71ac55c965

+++ autotest/cpp/test_cpl.cpp
@@ -153,7 +153,7 @@ namespace tut
         {
             ensure(CPLGetValueType(apszTestStrings[i].testString) == apszTestStrings[i].expectedResult);
             if (CPLGetValueType(apszTestStrings[i].testString) != apszTestStrings[i].expectedResult)
-                fprintf(stderr, "mismatch on item %d : value=\"%s\", expect_result=%d, result=%d\n", i,
+                fprintf(stderr, "mismatch on item %d : value=\"%s\", expect_result=%d, result=%d\n", (int)i,
                         apszTestStrings[i].testString, 
                         apszTestStrings[i].expectedResult,
                         CPLGetValueType(apszTestStrings[i].testString));
